lots of networking changes
- use dynamic netIds for units & players
- changed player packets to reference units instead for easier future use
- added packet to register & verify UDP connection upon joining
- reduce packet size again by using Int16 (instead of Int32)

NOTE: this implementation will probably only work for localhost at the time being.

+++ core/main/units/Unit.cs
@@ -26,7 +26,7 @@ namespace Casanova.core.main.units
 
     public class Unit : Node2D
     {
-        public short netId;
+        public int netId;
         public Body Body;
         public Node2D Content;
         private readonly Array<Particles> Particles = new Array<Particles>();
